Clean up actions, consolidate container components
